clean up password stretching component and message ids
