+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/AdvancedSecurity/LoginIpRestriction/template.js
@@ -19,18 +19,18 @@ const LoginIpRestriction = props => {
       <SettingSummary>
         <SettingHeader>
           <FormattedMessage
-            id='scenes.securitysettings.advancedsettings.loginiprestriction.title'
+            id='scenes.securitycenter.advanced.loginiprestriction.title'
             defaultMessage='Login IP Restriction'
           />
           <SettingStatus active={ipLockOn}>
             {ipLockOn ? (
               <FormattedMessage
-                id='scenes.securitysettings.advancedsettings.loginiprestriction.enabled'
+                id='scenes.securitycenter.advanced.loginiprestriction.enabled'
                 defaultMessage='Enabled'
               />
             ) : (
               <FormattedMessage
-                id='scenes.securitysettings.advancedsettings.loginiprestriction.disabled'
+                id='scenes.securitycenter.advanced.loginiprestriction.disabled'
                 defaultMessage='Disabled'
               />
             )}
@@ -38,15 +38,15 @@ const LoginIpRestriction = props => {
         </SettingHeader>
         <SettingDescription>
           <FormattedMessage
-            id='scenes.securitysettings.advancedsettings.loginiprestriction.description'
+            id='scenes.securitycenter.advanced.loginiprestriction.description.part1'
             defaultMessage='Only allow login from IP address in your IP whitelist.'
           />
           <FormattedMessage
-            id='scenes.securitysettings.advancedsettings.loginiprestriction.description2'
+            id='scenes.securitycenter.advanced.loginiprestriction.description.part2'
             defaultMessage='If you do not have a static IP address, this may lock you out of your wallet.'
           />
           <FormattedMessage
-            id='scenes.securitysettings.advancedsettings.loginiprestriction.description3'
+            id='scenes.securitycenter.advanced.loginiprestriction.description.part3'
             defaultMessage='If you have verified your email address, you will be notified of any suspicious login attempts.'
           />
         </SettingDescription>
@@ -56,12 +56,12 @@ const LoginIpRestriction = props => {
           <Button nature='primary' onClick={handleClick}>
             {ipLockOn ? (
               <FormattedMessage
-                id='scenes.securitysettings.advancedsettings.loginiprestriction.settings.disable'
+                id='scenes.securitycenter.advanced.loginiprestriction.disable'
                 defaultMessage='Disable'
               />
             ) : (
               <FormattedMessage
-                id='scenes.securitysettings.advancedsettings.loginiprestriction.settings.enable'
+                id='scenes.securitycenter.advanced.loginiprestriction.enable'
                 defaultMessage='Enable'
               />
             )}
