+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/AdvancedSecurity/PasswordStretching/index.js
@@ -1,11 +1,60 @@
 import React from 'react'
+import PropTypes from 'prop-types'
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
+import { formValueSelector } from 'redux-form'
 
-import PasswordStretching from './template.js'
+import { actions, selectors } from 'data'
+import PasswordStretching from './template'
 
 class PasswordStretchingContainer extends React.PureComponent {
+  state = { updateToggled: false }
+
+  onSubmit = () => {
+    const { passwordStretchingValue } = this.props
+    this.props.walletActions.updatePbkdf2Iterations(
+      Number(passwordStretchingValue)
+    )
+    this.handleToggle()
+  }
+
+  handleToggle = () => {
+    this.props.formActions.reset('settingPasswordStretching')
+    this.setState({
+      updateToggled: !this.state.updateToggled
+    })
+  }
+
   render () {
-    return <PasswordStretching {...this.props} />
+    return (
+      <PasswordStretching
+        onSubmit={this.onSubmit}
+        updateToggled={this.state.updateToggled}
+        handleToggle={this.handleToggle}
+        {...this.props}
+      />
+    )
   }
 }
 
-export default PasswordStretchingContainer
+const mapStateToProps = state => ({
+  passwordStretchingValue: formValueSelector('settingPasswordStretching')(
+    state,
+    'passwordStretching'
+  ),
+  currentStretch: selectors.core.wallet.getPbkdf2Iterations(state)
+})
+
+const mapDispatchToProps = dispatch => ({
+  walletActions: bindActionCreators(actions.wallet, dispatch),
+  formActions: bindActionCreators(actions.form, dispatch)
+})
+
+PasswordStretchingContainer.propTypes = {
+  passwordStretching: PropTypes.number
+}
+
+export default connect(
+  mapStateToProps,
+  mapDispatchToProps
+)(PasswordStretchingContainer)
