+++ packages/blockchain-wallet-v4-frontend/src/scenes/SecurityCenter/AdvancedSecurity/SecondPasswordWallet/index.js
@@ -2,7 +2,7 @@ import React from 'react'
 import { connect } from 'react-redux'
 
 import { selectors } from 'data'
-import SecondPassword from './template.js'
+import SecondPassword from './template'
 
 class SecondPasswordWalletContainer extends React.PureComponent {
   render () {
